partial additions made for SQL database migration and linking tweetdeck accounts

+++ api.py
@@ -11,9 +11,13 @@ import werkzeug
 import flask
 import random
 import multiprocessing as mp
-from main import verify
+from main import verify, getAccountSettingFromDB, logData
 import re
 from datetime import datetime
+import sqlite3
+
+con = sqlite3.connect("database.db", check_same_thread=False)
+cur = con.cursor()
 
 # Returns a list of filenames
 def getList(name):
@@ -49,12 +53,15 @@ def getAccounts(request, pHash):
     p = request.cookies.get("p")
     if verify(p, pHash, "getAccounts") == True:
 
-        tmp = os.listdir("./configs/")
+        res = cur.execute("SELECT name FROM accounts")
+        tmp = res.fetchall()
         out = []
 
         for i in tmp:
-            if i.startswith(request.headers["StartsWith"]):
-                out.append(i[:i.index(".")])
+            if i[0].startswith(request.headers["StartsWith"]):
+                out.append(i[0])
+
+        
 
         returnCode = json.dumps({"accounts": out})
         logInfo(request.headers, request.remote_addr, returnCode)
@@ -65,7 +72,7 @@ def getAccounts(request, pHash):
         return returnCode
 
 
-def newConfig(request, pHash):
+def newConfig(request, pHash, template):
     data = request.data
     try:
         p = request.cookies.get("p")
@@ -73,8 +80,29 @@ def newConfig(request, pHash):
         p = request.headers["Password"]
     if verify(p, pHash, "newConfig") == True:
         try:
-            with open(f"./configs/{request.headers['AccountName']}.yml", "wb") as conf:
-                conf.write(data)
+            timeAdded = time.time()
+            insert = (
+                    template["name"],
+                    template["proxy"],
+                    template["guest_id"],
+                    template["auth_token"],
+                    template["ct0"],
+                    template["gt"],
+                    template["twid"],
+                    template["user-agent"],
+                    template["kdt"],
+                    timeAdded,
+                    template["cookie"],
+                    template["hours"],
+                    template["range"],
+                    template["rm"],
+                    template["deactivate"]
+                )
+            
+
+            cur.execute("INSERT OR IGNORE INTO accounts VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", insert)
+            con.commit()
+
             returnCode = "OK"
             logInfo(request.headers, request.remote_addr, returnCode)
             return returnCode
@@ -91,7 +119,12 @@ def delConfig(request, pHash):
         p = request.headers["Password"]
     if verify(p, pHash, "newConfig") == True:
         try:
-            os.remove(f"./configs/{request.headers['AccountName']}.yml")
+            if os.path.exists(f"./configs/{request.headers['AccountName']}.yml"):
+                os.remove(f"./configs/{request.headers['AccountName']}.yml")
+
+            cur.execute("DELETE * FROM accounts WHERE name = ?", (request.headers['AccountName'], ))
+            con.commit()
+
             returnCode = "OK"
             logInfo(request.headers, request.remote_addr, returnCode)
             return returnCode
@@ -101,24 +134,63 @@ def delConfig(request, pHash):
             return returnCode
 
 
+def newTweetdeckConfig(request, pHash, template):
+    data = request.data
+    try:
+        p = request.cookies.get("p")
+    except:
+        p = request.headers["Password"]
+    if verify(p, pHash, "newConfig") == True:
+        try:
+            timeAdded = time.time()
+            insert = (
+                    template["name"],
+                    template["guest_id"],
+                    template["auth_token"],
+                    template["ct0"],
+                    template["gt"],
+                    template["twid"],
+                    template["user-agent"],
+                    template["kdt"],
+                    timeAdded,
+                    template["cookie"],
+                )
+            
+
+            cur.execute("INSERT OR IGNORE INTO tweetdeckAccounts VALUES(?, ?, ?, ?, ?, ?, ?, ?, ?, ?)", insert)
+            con.commit()
+
+            returnCode = "OK"
+            logInfo(request.headers, request.remote_addr, returnCode)
+            return returnCode
+        except:
+            returnCode = "ERROR MAKING CONFIG"
+            logInfo(request.headers, request.remote_addr, returnCode)
+            return returnCode
+
+
 def stop(request, pHash):
     p = request.cookies.get("p")
     if verify(p, pHash, "stop") == True:
         try:
-            with open(f"./configs/{request.headers['AccountName']}.yml", "r") as conf:
+            if os.path.exists("./configs/{request.headers['AccountName']}.yml"):
+                with open(f"./configs/{request.headers['AccountName']}.yml", "r") as conf:
+                    
+                    out = ""
+                    for i in conf.readlines():
+                        if i.startswith("deactivate: 0"):
+                            out += "deactivate: 1\n"
+                        else:
+                            out += f"{i}"
+                with open(f"./configs/{request.headers['AccountName']}.yml", "w") as conf:
+                    conf.write(out)
+
+            cur.execute(f"UPDATE accounts SET deactivate = ? WHERE name = ?", (1, request.headers['AccountName']))
+            con.commit()
                 
-                out = ""
-                for i in conf.readlines():
-                    if i.startswith("deactivate: 0"):
-                        out += "deactivate: 1\n"
-                    else:
-                        out += f"{i}"
-            with open(f"./configs/{request.headers['AccountName']}.yml", "w") as conf:
-                conf.write(out)
-                
-                returnCode = "OK"
-                logInfo(request.headers, request.remote_addr, returnCode)
-                return returnCode
+            returnCode = "OK"
+            logInfo(request.headers, request.remote_addr, returnCode)
+            return returnCode
         except:
             returnCode = "ERROR EDITING CONFIG"
             logInfo(request.headers, request.remote_addr, returnCode)
@@ -133,20 +205,24 @@ def restart(request, pHash):
     p = request.cookies.get("p")
     if verify(p, pHash, "restart") == True:
         try:
-            with open(f"./configs/{request.headers['AccountName']}.yml", "r") as conf:
-                out = ""
-                for i in conf.readlines():
-                    if i.startswith("deactivate: 1"):
-                        out += "deactivate: 0\n"
-                    else:
-                        out += f"{i}"
+            if os.path.exists(f"./configs/{request.headers['AccountName']}.yml"):
+                with open(f"./configs/{request.headers['AccountName']}.yml", "r") as conf:
+                    out = ""
+                    for i in conf.readlines():
+                        if i.startswith("deactivate: 1"):
+                            out += "deactivate: 0\n"
+                        else:
+                            out += f"{i}"
 
-            with open(f"./configs/{request.headers['AccountName']}.yml", "w") as conf:
-                conf.write(out)
+                with open(f"./configs/{request.headers['AccountName']}.yml", "w") as conf:
+                    conf.write(out)
+
+            cur.execute(f"UPDATE accounts SET deactivate = ? WHERE name = ?", (0, request.headers['AccountName']))
+            con.commit()
                 
-                returnCode = "OK"
-                logInfo(request.headers, request.remote_addr, returnCode)
-                return returnCode
+            returnCode = "OK"
+            logInfo(request.headers, request.remote_addr, returnCode)
+            return returnCode
         except:
             returnCode = "ERROR EDITING CONFIG"
             logInfo(request.headers, request.remote_addr, returnCode)
@@ -202,7 +278,6 @@ def getData(request, pHash, hours, deactivate):
 def likeTweet(
     tweet,
     proxy,
-    authorization,
     guest_id,
     ct0,
     kdt,
@@ -227,7 +302,7 @@ def likeTweet(
         "Accept-Encoding": "gzip, deflate, br",
         "Content-Type": "application/json",
         "Accept-Language": "en-US,en;q=0.5",
-        "authorization": f"Bearer {authorization}",
+        "authorization": f"Bearer AAAAAAAAAAAAAAAAAAAAANRILgAAAAAAnNwIzUejRCOuH5E6I8xnZz4puTs%3D1Zv7ttfk8LF81IUq16cHjhLTvJu4FA33AGWWjCpTnA",
         "Connection": "keep-alive",
         "Content-Length": f"{len(data)}",
         "Cookie": f"guest_id={guest_id}; ct0={ct0}; kdt={kdt}; twid={twid}; auth_token={auth_token}; gt={gt};",
@@ -253,12 +328,66 @@ def likeTweet(
         res = requests.post(url=url, headers=h, data=data, timeout=10)
         print(res.text)
 
+def tweetdeckLikeTweet(
+    tweet,
+    proxy,
+    guest_id,
+    ct0,
+    kdt,
+    twid,
+    auth_token,
+    gt,
+    userAgent,
+    userID
+):
 
+    url = "https://api.twitter.com/1.1/favorites/create.json"
+
+    data = f"id={tweet}&cards_platform=Web-13&include_entities=1&include_user_entities=1&include_cards=1&send_error_codes=1&tweet_mode=extended&include_ext_alt_text=true&include_reply_count=true"
+
+    if proxy != None:
+        p = {
+            "http": f"{proxy}"
+        }
+    else:
+        p = None
+    
+    h = {
+        "Accept": "*/*",
+        "Accept-Encoding": "gzip, deflate, br",
+        "Content-Type": "application/x-www-form-urlencoded; charset=UTF-8",
+        "Accept-Language": "en-US,en;q=0.5",
+        "authorization": f"Bearer AAAAAAAAAAAAAAAAAAAAANRILgAAAAAAnNwIzUejRCOuH5E6I8xnZz4puTs%3D1Zv7ttfk8LF81IUq16cHjhLTvJu4FA33AGWWjCpTnA",
+        "Connection": "keep-alive",
+        "Content-Length": f"{len(data)}",
+        "Cookie": f"guest_id={guest_id}; ct0={ct0}; kdt={kdt}; twid={twid}; auth_token={auth_token}; gt={template['gt']};",
+        "DNT": "1",
+        "host": "api.twitter.com",
+        "origin": "https://tweetdeck.twitter.com",
+        "referer": "https://tweetdeck.twitter.com/",
+        "sec-fetch-dest": "empty",
+        "sec-fetch-mode": "cors",
+        "sec-fetch-site": "same-site",
+        "TE": "trailers",
+        "User-Agent": f"{userAgent}",
+        "x-csrf-token": f"{ct0}",
+        "x-twitter-active-user": "yes",
+        "x-twitter-auth-type": "OAuth2Session",
+        "x-twitter-client-language": "en",
+        "x-act-as-user-id": f"{userID}",
+        "x-twitter-client-version": "Twitter-TweetDeck-blackbird-chrome/4.0.220811153004 web/"
+    }
+
+    if p != None:
+        res = requests.post(url=url, headers=h, data=data, timeout=10, proxies=p)
+        print(res.text)
+    else:
+        res = requests.post(url=url, headers=h, data=data, timeout=10)
+        print(res.text)
 
 def deleteTweet(
     request,
     proxy,
-    authorization,
     guest_id,
     ct0,
     kdt,
@@ -296,7 +425,7 @@ def deleteTweet(
                 "Accept-Encoding": "gzip, deflate, br",
                 "Content-Type": "application/json",
                 "Accept-Language": "en-US,en;q=0.5",
-                "authorization": f"Bearer {authorization}",
+                "authorization": f"Bearer AAAAAAAAAAAAAAAAAAAAANRILgAAAAAAnNwIzUejRCOuH5E6I8xnZz4puTs%3D1Zv7ttfk8LF81IUq16cHjhLTvJu4FA33AGWWjCpTnA",
                 "Connection": "keep-alive",
                 "Content-Length": f"{len(data)}",
                 "Cookie": f"guest_id={guest_id}; ct0={ct0}; kdt={kdt}; twid={twid}; auth_token={auth_token}; gt={gt};",
@@ -344,60 +473,64 @@ def changeSettings(request):
     try:
         data = json.loads(request.data)
 
-        valid = ["hours", "range", "delete", "proxy"]
+        valid = ["hours", "range", "rm", "proxy"]
 
         if contains_duplicates(list(data.keys())) == True:
             returnCode = "IMPROPER INPUT"
             logInfo(request.headers, request.remote_addr, returnCode)
             return returnCode
 
+        name = request.headers['AccountName']
+
+        hours = float(getAccountSettingFromDB(name, "hours"))
+        range = float(getAccountSettingFromDB(name, "range"))
+        rm = getAccountSettingFromDB(name, "rm")
+        proxy = getAccountSettingFromDB(name, "proxy")
+
         for i in data:
             print(data[i])
             print(i)
             if i in valid:
-                with open(f"./configs/{request.headers['AccountName']}.yml", "r") as conf:
-                    template = yaml.safe_load(conf)
+                if i == "rm" and int(data[i]) not in [0, 1]:
+                    returnCode = "IMPROPER INPUT"
+                    logInfo(request.headers, request.remote_addr, returnCode)
+                    return returnCode
 
-                    if i == "delete" and int(data[i]) not in [0, 1]:
-                        returnCode = "IMPROPER INPUT"
-                        logInfo(request.headers, request.remote_addr, returnCode)
-                        return returnCode
+                if i == "proxy" and len(str(data[i])) < 0:
+                    returnCode = "IMPROPER INPUT"
+                    logInfo(request.headers, request.remote_addr, returnCode)
+                    return returnCode
 
-                    if i == "proxy" and len(str(data[i])) < 0:
-                        returnCode = "IMPROPER INPUT"
-                        logInfo(request.headers, request.remote_addr, returnCode)
-                        return returnCode
-
-                    if i == "range" and int(data[i]) < 0:
-                        returnCode = "IMPROPER INPUT"
-                        logInfo(request.headers, request.remote_addr, returnCode)
-                        return returnCode
-                    elif i == "range" and int(data[i]) > template["hours"]:
-                        if "hours" in data:
-                            if int(data["hours"]) <= int(data[i]):
-                                returnCode = "IMPROPER INPUT"
-                                logInfo(request.headers, request.remote_addr, returnCode)
-                                return returnCode
-                        else:
+                if i == "range" and int(data[i]) < 0:
+                    returnCode = "IMPROPER INPUT"
+                    logInfo(request.headers, request.remote_addr, returnCode)
+                    return returnCode
+                elif i == "range" and int(data[i]) > hours:
+                    if "hours" in data:
+                        if int(data["hours"]) <= int(data[i]):
                             returnCode = "IMPROPER INPUT"
                             logInfo(request.headers, request.remote_addr, returnCode)
                             return returnCode
-
-
-                    if i == "hours" and int(data[i]) < 0:
+                    else:
                         returnCode = "IMPROPER INPUT"
                         logInfo(request.headers, request.remote_addr, returnCode)
                         return returnCode
-                    elif i == "hours" and int(data[i]) < template["range"]:
-                        if "range" in data:
-                            if int(data["range"]) >= int(data[i]):
-                                returnCode = "IMPROPER INPUT"
-                                logInfo(request.headers, request.remote_addr, returnCode)
-                                return returnCode
-                        else:
+
+
+                if i == "hours" and int(data[i]) < 0:
+                    returnCode = "IMPROPER INPUT"
+                    logInfo(request.headers, request.remote_addr, returnCode)
+                    return returnCode
+                elif i == "hours" and int(data[i]) < range:
+                    if "range" in data:
+                        if int(data["range"]) >= int(data[i]):
                             returnCode = "IMPROPER INPUT"
                             logInfo(request.headers, request.remote_addr, returnCode)
                             return returnCode
+                    else:
+                        returnCode = "IMPROPER INPUT"
+                        logInfo(request.headers, request.remote_addr, returnCode)
+                        return returnCode
             else:
                 returnCode = "IMPROPER INPUT"
                 logInfo(request.headers, request.remote_addr, returnCode)
@@ -405,18 +538,21 @@ def changeSettings(request):
 
         for i in data:
             if i in valid:
-                with open(f"./configs/{request.headers['AccountName']}.yml", "r") as conf:
-                    out = ""
-                    for j in conf.readlines():
-                        if j.startswith(f"{i}:"):
-                            out += f"{i}: {data[i]}\n"
-                        else:
-                            out += f"{j}"
+                if os.path.exists(f"./configs/{request.headers['AccountName']}.yml"):
+                    with open(f"./configs/{request.headers['AccountName']}.yml", "r") as conf:
+                        out = ""
+                        for j in conf.readlines():
+                            if j.startswith(f"{i}:"):
+                                out += f"{i}: {data[i]}\n"
+                            else:
+                                out += f"{j}"
 
+                    with open(f"./configs/{request.headers['AccountName']}.yml", "w") as conf:
+                        conf.write(out)
 
-                with open(f"./configs/{request.headers['AccountName']}.yml", "w") as conf:
-                    conf.write(out)
-                    
+                cur.execute(f"UPDATE accounts SET {i} = ? WHERE name = ?", (data[i], request.headers['AccountName']))
+                con.commit()
+
         returnCode = "OK"
         logInfo(request.headers, request.remote_addr, returnCode)
         return returnCode
